test: move null event to distpatcher test file
