+++ dispatcher_test.go
@@ -1,6 +1,8 @@
 package eventstore
 
 import (
+	"bytes"
+	"encoding/binary"
 	"errors"
 	"sync"
 	"testing"
@@ -29,6 +31,33 @@ func (n *slowReducer) Reduce(event Event) error {
 	time.Sleep(2 * time.Second)
 	return nil
 }
+
+type nullEvent struct {
+	Timestamp time.Time
+}
+
+func (n *nullEvent) Body() []byte {
+	return nil
+}
+
+func (n *nullEvent) Time() []byte {
+	t := n.Timestamp.UnixNano()
+	buf := new(bytes.Buffer)
+	// Use big endian to preserve lexicographic sorting
+	binary.Write(buf, binary.BigEndian, t)
+	return buf.Bytes()
+}
+
+func (n *nullEvent) EntityID() string {
+	return "null"
+}
+
+func (n *nullEvent) Type() string {
+	return "null"
+}
+
+// Sanity check
+var _ Event = (*nullEvent)(nil)
 
 func setup() *Dispatcher {
 	return NewDispatcher(datastore.NewMapDatastore())
